api-gateway/requirements.txt
# Data Acuity API Gateway - Python Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
starlette==0.36.3

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# YAML (for OpenAPI)
PyYAML==6.0.1

# Security
python-multipart==0.0.9

# Logging
structlog==24.1.0
